% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sql.R
\name{sql}
\alias{as.sql}
\alias{as.sql.character}
\alias{is.sql}
\alias{print.sql}
\alias{sql}
\title{Represent a character vector as a SQL statement}
\usage{
sql(x)

is.sql(x)

\method{print}{sql}(x, ...)

as.sql(x)

\method{as.sql}{character}(x)
}
\arguments{
\item{x}{character; string to convert to a SQL statement}

\item{...}{additional arguments}
}
\description{
Creates a string with class 'sql'
}
\details{
There is nothing magical here. An object of sql is just a string ... i.e.
an one-element character vector. It is coerced to a string by 'collapsing' the
vector with \code{\\n}.
}
\examples{
stmt = "SELECT * FROM table"

  sql(stmt)

  is.sql(stmt)        # FALSE
  is.sql( sql(stmt) ) # TRUE

  as.sql(stmt)
}
\seealso{
\code{\link{sql_read}} for reading SQL from a connection
}

