% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{sql.template}
\alias{sql.template}
\alias{sql.template-package}
\title{sql.template - lightweight SQL templating engine}
\description{
\strong{sql.template} is a simple, mustache-based SQL templating library. It
allows  writing syntactically valid SQL that can execute as-is and
simultaneously support injection of values programattically before sending
the query for execution. It can:

\itemize{
  \item substitution using \code{whisker} templates
  \item tags that allow for unmasking of SQL templating code
  \item read SQL from any connection-type object such as a file.
}

This is accomplished using the \strong{\code{\link{sql_render}}} function.

The package was designed so that the developers could share SQL code with
other developers from a shared or forked repository.
}
\examples{
sql = "
SELECT
  *
FROM
  {{table}}
WHERE
  1 == 1
--r: FIRST_NAME = '{{first_name}}'
"

sql <- sql_render(
    sql
  , list( table = "table_1", first_name = "Barack" )
  , render=TRUE
)

cat( sql )
}
\references{
\url{http://en.wikipedia.org/wiki/Mustache_(template_system)} \cr
  \url{http://mustache.github} \cr
  \url{https://github.com/edwindj/whisker}
}
\seealso{
\link{sql_render}
}

