% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sql_render}
\alias{sql_render}
\title{Render SQL template}
\usage{
sql_render(sql, data = parent.frame(), tags = "r", strip.comments = TRUE,
  render = TRUE)
}
\arguments{
\item{sql}{connection object or character; Path to \code{SQL} file.}

\item{data}{environment, list or named vector. Locations to search for values
to substitute into the query. Default: parent.frame}

\item{tags}{character; list of tags to uncomment.  Default: "r"}

\item{strip.comments}{logical; whether to strip comments from sql;
Default: TRUE

Reads in a SQL statements, renders it using \code{whisker}.

Strips tags, renders variables and returns the variable

\code{sql_render} makes no attempt to determine if resulting SQL is valid SQL.}

\item{render}{logical; whether to make a template replacement using
\code{whisker.render} (DEFAULT:TRUE)}
}
\value{
character; a SQL statement that can then be passed to a DATABASE
connection
}
\description{
Read SQL and substitute parameters
}
\examples{
sql = "select * from {{table}}"
  sql_render( sql )

  table = "table_1"
  sql_render( sql )
  sql_render( "select * from {{table}} --r: where ROWNUM < 10 ")

  # with magrittr:
  \dontrun{
    sql \%>\% sql_render( c( table = "table4" ) )
  }
}
\seealso{
\code{\link[whisker]{whisker.render}} \cr
}

