% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sql_strip_comments.R
\docType{data}
\name{sql_strip_comments}
\alias{sql_strip_comments}
\title{sql_strip_comments}
\format{\preformatted{ NULL
}}
\usage{
sql_strip_comments
}
\arguments{
\item{sql}{connection or character;}
}
\description{
Strip comments from a SQL STATEMENT
}
\details{
#

This function has not been written yet.
}
\keyword{datasets}

