% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sql_render.R
\name{sql_render}
\alias{sql_render}
\title{Render SQL template}
\usage{
sql_render(sql, data = NULL, tags = "r", strip.comments = TRUE,
  render = TRUE)
}
\arguments{
\item{sql}{character or connection object; Can be path to \code{SQL} file,
raw SQL or a connection object.}

\item{data}{environment, list or named vector. Locations to search for values
to substitute into the query. The default, \code{NULL} finds the names from
the call stack.}

\item{tags}{character; list of tags to uncomment.  Default: "r"}

\item{strip.comments}{logical; whether to strip comments from sql;
Default: TRUE

Reads in a SQL statements, renders it using \code{whisker}. If \code{sql} is
character that matches a file on the file system this is read, otherwise
\code{character} is taken a a literal statement. \code{sql} can also be a
connection object.

Strips tags, renders variables and returns the variable

\code{sql_render} makes no attempt to determine if resulting SQL is valid SQL.}

\item{render}{logical; whether to make a template replacement using
\code{whisker.render} (DEFAULT:TRUE)}
}
\value{
character; a SQL statement that can then be passed to a DATABASE
connection
}
\description{
Read SQL and substitute parameters
}
\examples{
sql = "select * from {{table}}"
  table = "table1"
  sql_render( sql )

  table = "table_1"
  sql_render( sql )
  sql_render( "select * from {{table}} --r: where ROWNUM < 10 ")

  # with magrittr:
  \dontrun{
    sql \%>\% sql_render( c( table = "table4" ) )
  }
}
\seealso{
\code{\link[whisker]{whisker.render}} \cr
}

